requirements.txt
# Core
# gradio 5+ changed Chatbot default to type="messages"; app.py uses type="tuples" explicitly
gradio==6.12.0
# ollama 0.6+ changed embed() / list() / chat() to return Pydantic objects
ollama==0.6.1
# chromadb 1.x — developed and tested against this version
chromadb==1.5.8
chardet==7.4.0.post2

# File extraction
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5

# Optional — cloud sync (install only what you use)
boto3==1.42.49                        # S3
google-api-python-client==2.194.0     # Google Drive
google-auth==2.49.2
google-auth-oauthlib==1.3.1
